+++ app/Http/Controllers/Admin/TreeView/IndikatorController.php
@@ -81,21 +81,27 @@ class IndikatorController extends Controller
             'nama_menu' => ['required', 'string']
         ]);
 
-        $table->update($request->all());
+        if ($table->id != 1) {
+            $table->update($request->all());
+        }
 
         return back();
     }
 
     public function destroy(TabelIndikator $table)
     {
-        $table->delete();
+        if ($table->id != 1) {
+            $table->delete();
+        }
 
         return back();
     }
 
     public function massDestroy(Request $request)
     {
-        TabelIndikator::whereIn('id', $request->ids)->delete();
+        $ids = collect($request->ids)->filter(fn ($val, $key) => $val != 1)->toArray();
+
+        TabelIndikator::whereIn('id', $ids)->delete();
 
         return response(null, Response::HTTP_NO_CONTENT);
     }
