+++ app/Http/Controllers/Admin/TreeView/RpjmdController.php
@@ -86,21 +86,27 @@ class RpjmdController extends Controller
             'nama_menu' => ['required', 'string',]
         ]);
 
-        $table->update($request->all());
+        if ($table->id != 1) {
+            $table->update($request->all());
+        }
 
         return back();
     }
 
     public function destroy(TabelRpjmd $table)
     {
-        $table->delete();
+        if ($table->id != 1) {
+            $table->delete();
+        }
 
         return back();
     }
 
     public function massDestroy(Request $request)
     {
-        TabelRpjmd::whereIn('id', $request->ids)->delete();
+        $ids = collect($request->ids)->filter(fn ($val, $key) => $val != 1)->toArray();
+
+        TabelRpjmd::whereIn('id', $ids)->delete();
 
         return response(null, Response::HTTP_NO_CONTENT);
     }
